Added support to import csv files (does not work yet) Added entities for the project an its optimization outputs that will be imported from csv file.
Added views to upload the csv file and save as as a project, edit and delete

Added IOptimizationManager, with implementation OptimizationManager. and registered in the DI. The idea is to inject this interface instead of the datacontext to each PageModel

Problems with applying migrations from Dockerfile. Didn't work. Also the upload file page fails when uploading the file because of "bad data" raised by CvsHelper. Also I have to manage a bit more the library to parse and convert colums dynamically.